pyproject.toml
[project]
name = "acb-chatbot-dev-program"
version = "0.1.0"
description = "This project contains code relevant to the ACB chatbot development program. "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pathlib>=1.0.1",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
]
